fix custom init data
