--- lib/tasks/turnip_support_init_data.rake
@@ -1,6 +1,0 @@
-namespace :initialization do
-  task init_test_data: :environment do
-    puts "You can put the data initialization for feature test in #{Rails.root}/lib/tasks/turnip_support_init_data.rake"
-    # TODO
-  end
-end
