+++ spec/steps/custom_steps.rb
@@ -1,0 +1,4 @@
+step "custom_init_data" do
+  # TODO
+  puts "You can customize initializing data in the spec/steps/custom_steps.rb"
+end
